fix bug in C++ calc_min_max example (#134)
By starting the reduction with a (0, 0) tuple, the function returned
an incorrect minimum of zero if all inputs were positive or an incorrect
maximum of zero if all inputs were negative.

The new version assumes that the `data` vector is not empty, which is
always the case here but might not be in real life. I think that's fine
for an example.

I also swapped `std::reduce` for `std::accumulate` to be more
consistent with the rest of the text. It's sequential by default,
and we're still providing an initial element, so I think it's
strictly equivalent in this case, but I haven't tested it.